make findDomainIdFromOperatorsCache utils function and implement it
