+++ indexers/staking/src/mappings/utils.ts
@@ -1,6 +1,7 @@
 import { EventRecord, stringify } from "@autonomys/auto-utils";
 import { createHash } from "crypto";
 import { PAD_ZEROS, ZERO_BIGINT } from "./constants";
+import { Cache } from "./db";
 import { Transfer } from "./types";
 
 export const getSortId = (
@@ -43,4 +44,21 @@ export const findOneExtrinsicEvent = (
       e.event.section === section &&
       e.event.method === method
   );
+};
+
+export const findDomainIdFromOperatorsCache = (
+  cache: Cache,
+  operatorId: string
+): string => {
+  const opFromCache = cache.operatorStakingHistory.find(
+    (o) => o.operatorId === operatorId
+  );
+  if (!opFromCache) {
+    const parentOpFromCache = cache.parentBlockOperators.find(
+      (o) => o.operatorId.toString() === operatorId
+    );
+    if (!parentOpFromCache) throw new Error("Operator from cache not found");
+    return parentOpFromCache.operatorDetails.currentDomainId.toString();
+  }
+  return opFromCache.currentDomainId;
 };
